Add librusqlite_noicu
librusqlite_noicu is a variant of librusqlite which is linked to
libsqlite built with no ICU extension support. This variant is intended
to be used in environments where the ICU extension is not needed and the
dependency to ICU (which is as large as 60+ MB) is not desirable, like
microdroid.

Bug: 199674764
Test: m
Change-Id: I7a1412baa449606939dd291f51bccc871f306ec1